+++ react/components/drawer/ComparisonDrawer.tsx
@@ -1,18 +1,20 @@
 import React, { useMemo } from 'react'
-import { pathOr, uniq, find, propEq } from 'ramda'
+import { pathOr, uniq, find, propEq, isEmpty } from 'ramda'
 import ComparisonContext from '../../ProductComparisonContext'
 import productsQuery from '../../queries/productsByIdentifier.graphql'
 import { useQuery } from 'react-apollo'
 import styles from './drawer.css'
+import { Button } from 'vtex.styleguide'
+import ProductThumbnail from './ProductThumbnail'
 
 const ComparisonDrawer = () => {
   const {
     useProductComparisonState,
-    // useProductComparisonDispatch,
+    useProductComparisonDispatch,
   } = ComparisonContext
 
   const comparisonData = useProductComparisonState()
-  // const dispatchComparison = useProductComparisonDispatch()
+  const dispatchComparison = useProductComparisonDispatch()
 
   const comparisonProducts = pathOr(
     [] as ProductToCompare[],
@@ -39,6 +41,8 @@ const ComparisonDrawer = () => {
             propEq('itemId', productInfo.skuId.toString())
           )(pathOr([], ['items'], selectedProduct))
           return {
+            productId: pathOr('', ['productId'], selectedProduct),
+            skuId: pathOr('', ['itemId'], selectedSku),
             imageUrl: pathOr('', ['images', 0, 'imageUrl'], selectedSku),
             productName: pathOr('', ['productName'], selectedProduct),
             skuName: pathOr('', ['name'], selectedSku),
@@ -46,32 +50,51 @@ const ComparisonDrawer = () => {
         })
   }, [comparisonProducts, productsResponse])
 
+  const removeAllItems = () => {
+    dispatchComparison({
+      type: 'REMOVE_ALL',
+    })
+  }
+
   return loading ? (
     <div>Loading...</div>
   ) : error ? (
     <div>{error}</div>
+  ) : isEmpty(thumbnailItems) ? (
+    <div />
   ) : (
     <div className={`${styles.drawerContainer} mw9 w-100 flex justify-center`}>
-      <div className={`${styles.drawer} flex flex-row justify-center`}>
+      <div className={`${styles.drawer} flex flex-row justify-center pl3 pr3`}>
         {thumbnailItems.map(thumbnail => {
           return (
-            <div
-              className="ma3"
-              key={`${thumbnail.productName}-${thumbnail.skuName}`}
-            >
-              <div className={`${styles.drawerImageContainer} pa3`}>
-                <img
-                  className={`${styles.drawerImage}`}
-                  alt={thumbnail.productName}
-                  src={thumbnail.imageUrl}
-                />
-              </div>
-              <div>
-                <span>{thumbnail.skuName}</span>
-              </div>
-            </div>
+            <ProductThumbnail
+              key={`${thumbnail.productId}-${thumbnail.skuId}`}
+              productThumbnail={thumbnail}
+            />
           )
         })}
+        <div className={`${styles.comparisonButtons} flex flex-column ma3`}>
+          <div className="">
+            <Button
+              block
+              size="small"
+              className={`${styles.compareProductsButton} ma3`}
+              href="/compare"
+            >
+              Compare
+            </Button>
+          </div>
+          <div className="mt3">
+            <Button
+              block
+              variation="danger-tertiary"
+              size="small"
+              onClick={removeAllItems}
+            >
+              Remove All
+            </Button>
+          </div>
+        </div>
       </div>
     </div>
   )